+++ src/views/MainViewProvider.ts
@@ -2,16 +2,19 @@ import * as vscode from 'vscode';
 import * as fs from 'fs';
 import { ConfigFile } from '../classes/configFile';
 import { RequestModel } from '../classes/requestModel';
+import { ChatSessions } from '../classes/chatSessions';
 import { updateConfigurations } from '../utils/configuration';
 
 export class MainViewProvider implements vscode.WebviewViewProvider {
     public static readonly viewType = 'light-assistant.main';
     private _view?: vscode.WebviewView;
+    private config = vscode.workspace.getConfiguration('lightAssistant');
     constructor(
         private readonly _extensionUri: vscode.Uri,
         private faIcons: any,
         private configFile: ConfigFile,
-        private requestModel: RequestModel
+        private requestModel: RequestModel,
+        private chatSessions: ChatSessions
     ) {}
 
     public updateConfiguration() {
@@ -19,7 +22,23 @@ export class MainViewProvider implements vscode.WebviewViewProvider {
     }
 
     public newChatSession() {
-        this.requestModel.newChatSession(this._view);
+        this.chatSessions.newChatSession(this._view);
+    }
+
+    public loadChatSession(fileName: string) {
+        this.chatSessions.loadChatSession(fileName, this._view);
+    }
+
+    public deleteChatSession(fileName: string) {
+        this.chatSessions.deleteChatSession(fileName, this._view);
+    }
+    public initView(){
+        this._view?.webview.postMessage({command: 'icons', icons: JSON.stringify(this.faIcons)});
+        this.updateConfiguration();
+        this.configFile.updateModelListFromConfig(this._view);
+        if(this.config.get<boolean>('loadLastChatSession')){
+            this.chatSessions.loadLastChatSession(this._view);
+        }
     }
 
     public resolveWebviewView(
@@ -35,12 +54,10 @@ export class MainViewProvider implements vscode.WebviewViewProvider {
         webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
 
         webviewView.webview.onDidReceiveMessage(message => {
-            console.log(message);
+            // console.log(message);
             switch (message.command) {
                 case 'init.ready':
-                    this._view?.webview.postMessage({command: 'icons', icons: JSON.stringify(this.faIcons)});
-                    this.updateConfiguration();
-                    this.configFile.updateModelListFromConfig(this._view);
+                    this.initView();
                     break;
                 case 'error.noModel':
                     vscode.window.showErrorMessage("No model selected, please select a model first.");
@@ -57,6 +74,9 @@ export class MainViewProvider implements vscode.WebviewViewProvider {
                     break;
                 case 'model.add':
                     this.configFile.addModelToConfig(message.modelData, this._view);
+                    break;
+                case 'model.delete':
+                    this.configFile.deleteModelFromConfig(message.modelData, this._view);
                     break;
             }
         });